requirements.txt
SQLAlchemy==1.1.11
ebi-ols-client==1.0.0